feat: remove project page
