--- app/dashboard/project/_actions/index.ts
@@ -1,8 +1,0 @@
-export { editProject } from "./edit-project.actions";
-export type { EditProjectResult } from "./edit-project.actions";
-export { createProject } from "./create-project.actions";
-export type { CreateProjectResult } from "./create-project.actions";
-export { joinProject } from "./join-project.actions";
-export type { JoinProjectResult } from "./join-project.actions";
-export { leaveProject } from "./leave-project.actions";
-export type { LeaveProjectResult } from "./leave-project.actions";
