--- app/dashboard/project/_components/_empty-project/create-project-dialog.tsx
@@ -1,118 +1,0 @@
-"use client";
-
-import { Loader2 } from "lucide-react";
-import { Controller } from "react-hook-form";
-
-import { useCreateProjectDialog } from "@/app/dashboard/project/_hooks";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { Field, FieldError, FieldGroup, FieldLabel, FieldSet } from "@/components/ui/field";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-
-const CreateProjectDialog = () => {
-  const { control, handleSubmit, onSubmit, isSubmitting, isOpen, setIsOpen } = useCreateProjectDialog();
-
-  return (
-    <Dialog open={isOpen} onOpenChange={setIsOpen}>
-      <DialogTrigger asChild>
-        <Button>Create new project</Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-[400px] md:max-w-[600px] lg:max-w-[800px]">
-        <DialogHeader>
-          <DialogTitle>Create new project</DialogTitle>
-          <DialogDescription>
-            Please enter your project details to continue. These can be changed later.
-          </DialogDescription>
-        </DialogHeader>
-        <form id="create-project-form" onSubmit={handleSubmit(onSubmit)}>
-          <FieldGroup>
-            <FieldSet disabled={isSubmitting}>
-              <Controller
-                name="name"
-                control={control}
-                render={({ field, fieldState }) => (
-                  <Field data-invalid={fieldState.invalid}>
-                    <FieldLabel htmlFor={field.name} className="after:content-['*'] after:ml-0.5 after:text-red-500">
-                      Project Name
-                    </FieldLabel>
-                    <Input
-                      {...field}
-                      id={field.name}
-                      placeholder="Enter your project name"
-                      aria-invalid={fieldState.invalid}
-                      autoComplete="off"
-                    />
-                    {fieldState.invalid && <FieldError errors={[fieldState.error]} className="w-full text-start" />}
-                  </Field>
-                )}
-              />
-            </FieldSet>
-            <FieldSet disabled={isSubmitting}>
-              <Controller
-                name="description"
-                control={control}
-                render={({ field, fieldState }) => (
-                  <Field data-invalid={fieldState.invalid}>
-                    <FieldLabel htmlFor={field.name} className="after:content-['*'] after:ml-0.5 after:text-red-500">
-                      Project Description
-                    </FieldLabel>
-                    <Textarea
-                      {...field}
-                      id={field.name}
-                      placeholder="Enter your project description"
-                      aria-invalid={fieldState.invalid}
-                      autoComplete="off"
-                      className="min-h-32 max-h-64 break-all"
-                    />
-                    {fieldState.invalid && <FieldError errors={[fieldState.error]} className="w-full text-start" />}
-                  </Field>
-                )}
-              />
-            </FieldSet>
-            <FieldSet disabled={isSubmitting}>
-              <Controller
-                name="github_url"
-                control={control}
-                render={({ field, fieldState }) => (
-                  <Field data-invalid={fieldState.invalid}>
-                    <FieldLabel htmlFor={field.name}>Github Repository URL</FieldLabel>
-                    <Input
-                      {...field}
-                      id={field.name}
-                      placeholder="Enter your Github repository URL"
-                      aria-invalid={fieldState.invalid}
-                      autoComplete="off"
-                    />
-                    {fieldState.invalid && <FieldError errors={[fieldState.error]} className="w-full text-start" />}
-                  </Field>
-                )}
-              />
-            </FieldSet>
-          </FieldGroup>
-        </form>
-        <DialogFooter>
-          <DialogClose asChild>
-            <Button type="button" variant="outline" disabled={isSubmitting}>
-              Go back
-            </Button>
-          </DialogClose>
-          <Button type="submit" form="create-project-form" disabled={isSubmitting}>
-            {isSubmitting ? <Loader2 className="animate-spin" /> : "Create project"}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export default CreateProjectDialog;
