--- app/dashboard/project/_components/index.ts
@@ -1,7 +1,0 @@
-export { default as CreateProjectDialog } from "./_empty-project/create-project-dialog";
-export { default as EditProjectForm } from "./edit-project-form";
-export { default as EmptyProject } from "./_empty-project/empty-project";
-export { default as JoinProjectDialog } from "./_empty-project/join-project-dialog";
-export { default as LeaveProjectDialog } from "./_team-members/leave-project-dialog";
-export { default as TeamMemberItem } from "./_team-members/team-member-item";
-export { default as TeamMembersList } from "./_team-members/team-members-list";
