--- app/dashboard/project/_schemas/edit-project-form.schemas.ts
@@ -1,11 +1,0 @@
-import { z } from "zod";
-
-import { secureUrlSchema } from "@/lib";
-
-import { createProjectFormSchema } from "./create-project-form.schemas";
-
-export const editProjectFormSchema = createProjectFormSchema.extend({
-  demo_url: secureUrlSchema.optional().or(z.literal("")),
-});
-
-export type EditProjectFormSchema = z.infer<typeof editProjectFormSchema>;
