--- app/dashboard/project/_schemas/index.ts
@@ -1,6 +1,0 @@
-export { createProjectFormSchema } from "./create-project-form.schemas";
-export type { CreateProjectFormSchema } from "./create-project-form.schemas";
-export { editProjectFormSchema } from "./edit-project-form.schemas";
-export type { EditProjectFormSchema } from "./edit-project-form.schemas";
-export { joinProjectFormSchema } from "./join-project-form.schemas";
-export type { JoinProjectFormSchema } from "./join-project-form.schemas";
