--- app/dashboard/project/_schemas/join-project-form.schemas.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const joinProjectFormSchema = z.object({
-  invitation_code: z.string().regex(/^[a-zA-Z0-9]{20}$/, "Invalid invitation code"),
-});
-
-export type JoinProjectFormSchema = z.infer<typeof joinProjectFormSchema>;
