--- app/dashboard/project/loading.tsx
@@ -1,12 +1,0 @@
-import { Skeleton } from "@/components/ui/skeleton";
-
-const ProjectLoading = async () => {
-  return (
-    <div className="flex-1 flex flex-col lg:flex-row lg:items-start gap-4">
-      <Skeleton className="flex-1 min-h-[525px]" />
-      <Skeleton className="flex-1 min-h-[500px]" />
-    </div>
-  );
-};
-
-export default ProjectLoading;
