+++ constants/index.ts
@@ -13,7 +13,6 @@ export {
 export {
   DASHBOARD_MANAGE_USERS_PATH,
   DASHBOARD_PATH,
-  DASHBOARD_PROJECT_PATH,
   DASHBOARD_SCHEDULE_PATH,
   DASHBOARD_SETTINGS_PATH,
   DASHBOARD_SUPPORT_PATH,
